move utils around
