+++ utils/functions.py
@@ -38,6 +38,7 @@ def get_positivity(string):
         return None
 
 
+# ==== search / select menus ====
 def search(list_to_search: list, value, key, cutoff=5, return_key=False, strict=False):
     """Fuzzy searches a list for an object
     result can be either an object or list of objects
@@ -155,18 +156,6 @@ async def search_and_select(ctx, list_to_search: list, query, key, cutoff=5, ret
     return result, metadata
 
 
-def a_or_an(string, upper=False):
-    if string.startswith('^') or string.endswith('^'):
-        return string.strip('^')
-    if re.match('[AEIOUaeiou].*', string):
-        return 'an {0}'.format(string) if not upper else f'An {string}'
-    return 'a {0}'.format(string) if not upper else f'A {string}'
-
-
-def camel_to_title(string):
-    return re.sub(r'((?<=[a-z])[A-Z]|(?<!\A)[A-Z](?=[a-z]))', r' \1', string).title()
-
-
 def paginate(iterable, n, fillvalue=None):
     args = [iter(iterable)] * n
     return [i for i in zip_longest(*args, fillvalue=fillvalue) if i is not None]
@@ -252,14 +241,6 @@ async def get_selection(ctx, choices, delete=True, pm=False, message=None, force
     return choices[int(m.content) - 1][1]
 
 
-ABILITY_MAP = {'str': 'Strength', 'dex': 'Dexterity', 'con': 'Constitution',
-               'int': 'Intelligence', 'wis': 'Wisdom', 'cha': 'Charisma'}
-
-
-def verbose_stat(stat):
-    return ABILITY_MAP[stat.lower()]
-
-
 async def confirm(ctx, message, delete_msgs=False):
     """
     Confirms whether a user wants to take an action.
@@ -284,6 +265,99 @@ async def confirm(ctx, message, delete_msgs=False):
     return replyBool
 
 
+# ==== display helpers ====
+def a_or_an(string, upper=False):
+    if string.startswith('^') or string.endswith('^'):
+        return string.strip('^')
+    if re.match('[AEIOUaeiou].*', string):
+        return 'an {0}'.format(string) if not upper else f'An {string}'
+    return 'a {0}'.format(string) if not upper else f'A {string}'
+
+
+def camel_to_title(string):
+    return re.sub(r'((?<=[a-z])[A-Z]|(?<!\A)[A-Z](?=[a-z]))', r' \1', string).title()
+
+
+def bubble_format(value: int, max_: int, fill_from_right=False):
+    """Returns a bubble string to represent a counter's value."""
+    if max_ > 100:
+        return f"{value}/{max_}"
+
+    used = max_ - value
+    filled = '\u25c9' * value
+    empty = '\u3007' * used
+    if fill_from_right:
+        return f"{empty}{filled}"
+    return f"{filled}{empty}"
+
+
+def verbose_stat(stat):
+    """Returns the long stat name for a abbreviation (e.g. "str" -> "Strength", etc)"""
+    return constants.STAT_ABBR_MAP[stat.lower()]
+
+
+def long_source_name(source):
+    return constants.SOURCE_MAP.get(source, source)
+
+
+def source_slug(source):
+    return constants.SOURCE_SLUG_MAP.get(source)
+
+
+def natural_join(things, between: str):
+    if len(things) < 3:
+        return f" {between} ".join(things)
+    first_part = ", ".join(things[:-1])
+    return f"{first_part}, {between} {things[-1]}"
+
+
+def trim_str(text, max_len):
+    """Trims a string to max_len."""
+    if len(text) < max_len:
+        return text
+    return f"{text[:max_len - 4]}..."
+
+
+def chunk_text(text, max_chunk_size=1024, chunk_on=('\n\n', '\n', '. ', ' '), chunker_i=0):
+    """
+    Recursively chunks *text* into a list of str, with each element no longer than *max_chunk_size*.
+    Prefers splitting on the elements of *chunk_on*, in order.
+    """
+
+    if len(text) <= max_chunk_size:  # the chunk is small enough
+        return [text]
+    if chunker_i >= len(chunk_on):  # we have no more preferred chunk_on characters
+        # optimization: instead of merging a thousand characters, just use list slicing
+        return [text[:max_chunk_size],
+                *chunk_text(text[max_chunk_size:], max_chunk_size, chunk_on, chunker_i + 1)]
+
+    # split on the current character
+    chunks = []
+    split_char = chunk_on[chunker_i]
+    for chunk in text.split(split_char):
+        chunk = f"{chunk}{split_char}"
+        if len(chunk) > max_chunk_size:  # this chunk needs to be split more, recurse
+            chunks.extend(chunk_text(chunk, max_chunk_size, chunk_on, chunker_i + 1))
+        elif chunks and len(chunk) + len(chunks[-1]) <= max_chunk_size:  # this chunk can be merged
+            chunks[-1] += chunk
+        else:
+            chunks.append(chunk)
+
+    # remove extra split_char from last chunk
+    chunks[-1] = chunks[-1][:-len(split_char)]
+    return chunks
+
+
+def smart_trim(text, max_len=1024):
+    """Uses chunk_text to return a trimmed str."""
+    chunks = chunk_text(text, max_len - 5)
+    out = chunks[0].strip()
+    if len(chunks) > 2:
+        return f"{chunks[0]}[...]"
+    return out
+
+
+# ==== misc helpers ====
 def auth_and_chan(ctx):
     """Message check: same author and channel"""
 
@@ -318,41 +392,7 @@ def maybe_mod(val: str, base=0):
     return base
 
 
-def bubble_format(value: int, max_: int, fill_from_right=False):
-    """Returns a bubble string to represent a counter's value."""
-    if max_ > 100:
-        return f"{value}/{max_}"
-
-    used = max_ - value
-    filled = '\u25c9' * value
-    empty = '\u3007' * used
-    if fill_from_right:
-        return f"{empty}{filled}"
-    return f"{filled}{empty}"
-
-
-def long_source_name(source):
-    return constants.SOURCE_MAP.get(source, source)
-
-
-def source_slug(source):
-    return constants.SOURCE_SLUG_MAP.get(source)
-
-
-def natural_join(things, between: str):
-    if len(things) < 3:
-        return f" {between} ".join(things)
-    first_part = ", ".join(things[:-1])
-    return f"{first_part}, {between} {things[-1]}"
-
-
-def trim_str(text, max_len):
-    """Trims a string to max_len."""
-    if len(text) < max_len:
-        return text
-    return f"{text[:max_len - 4]}..."
-
-
+# ==== user stuff ====
 async def user_from_id(ctx, the_id):
     """
     Gets a :class:`discord.User` given their user id in the context. Returns member if context has data.
